refactor repository definition to inject dependencies using hilt provided modules
