+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
         android:name="android.permission.INTERNET" />
 
     <application
-        android:name=".app.ComposeNewsApplication"
+        android:name=".ComposeNewsApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
